requirements.txt
gtrending==0.3.5
gitpython==3.1.24
pip-check-reqs==2.3.2
idna==3.3
packaging==21.2
Flask==2.0.2
